Homework3/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9


openai-whisper==20231117


transformers==4.43.3
accelerate==0.33.0
torch>=2.1.0


pyttsx3==2.90


pydantic==2.7.3
